Crypto Checker
This Python script allows you to retrieve the current price of a cryptocurrency, either just once or continuously at a specified interval. Simply specify the cryptocurrency of interest and the script will use real-time data from online exchanges to provide you with the most up-to-date price. Whether you are a cryptocurrency trader or just interested in keeping track of the market, this script is a quick and convenient way to stay informed.